+++ tests/testthat/test-connector.R
@@ -39,6 +39,21 @@ cli::test_that_cli("Test connector creation", {
   })
 })
 
+cli::test_that_cli("can create Connector object", {
+  temp_dir <- withr::local_tempdir("connector_test")
+
+  withr::with_tempdir(tmpdir = temp_dir, {
+    test <- ConnectorFS$new(path = ".")
+
+    connector_obj <- connectors(
+      test = test,
+      test_2 = base::as.data.frame(x = iris)
+    )
+
+    expect_snapshot_out(print(datasources(connector_obj)))
+  })
+})
+
 test_that("validate_resource works correctly", {
   temp_dir <- withr::local_tempdir("validate_test")